+++ pennyway-app-external-api/src/main/java/kr/co/pennyway/api/apis/test/TestDto.java
@@ -1,0 +1,5 @@
+package kr.co.pennyway.api.apis.test;
+
+public class TestDto {
+    private String name;
+}
