+++ pennyway-app-external-api/src/main/java/kr/co/pennyway/api/common/annotation/AccessTokenStrategy.java
@@ -1,0 +1,13 @@
+package kr.co.pennyway.api.common.annotation;
+
+import org.springframework.beans.factory.annotation.Qualifier;
+
+import java.lang.annotation.*;
+
+@Target({ElementType.FIELD, ElementType.PARAMETER, ElementType.METHOD,
+        ElementType.TYPE, ElementType.ANNOTATION_TYPE})
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+@Qualifier("accessTokenStrategy")
+public @interface AccessTokenStrategy {
+}
