+++ pennyway-app-external-api/src/main/java/kr/co/pennyway/api/common/security/jwt/refresh/RefreshTokenClaimKeys.java
@@ -1,0 +1,16 @@
+package kr.co.pennyway.api.common.security.jwt.refresh;
+
+public enum RefreshTokenClaimKeys {
+    USER_ID("userId"),
+    ROLE("role");
+
+    private final String value;
+
+    RefreshTokenClaimKeys(String value) {
+        this.value = value;
+    }
+
+    public String getValue() {
+        return value;
+    }
+}
