+++ pennyway-app-external-api/src/test/java/kr/co/pennyway/common/security/jwt/access/AccessTokenProviderTest.java
@@ -4,6 +4,8 @@ import kr.co.infra.common.exception.JwtErrorCode;
 import kr.co.infra.common.exception.JwtErrorException;
 import kr.co.infra.common.jwt.JwtClaims;
 import kr.co.infra.common.jwt.JwtProvider;
+import kr.co.pennyway.api.common.security.jwt.access.AccessTokenClaim;
+import kr.co.pennyway.api.common.security.jwt.access.AccessTokenProvider;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
