--- pennyway-domain/src/main/java/kr/co/pennyway/DomainPackageLocation.java
@@ -1,4 +1,0 @@
-package kr.co.pennyway;
-
-public interface DomainPackageLocation {
-}
