--- pennyway-domain/src/main/java/kr/co/pennyway/common/annotation/DomainRedisConnectionFactory.java
@@ -1,13 +1,0 @@
-package kr.co.pennyway.common.annotation;
-
-import org.springframework.beans.factory.annotation.Qualifier;
-
-import java.lang.annotation.*;
-
-@Target({ElementType.FIELD, ElementType.PARAMETER, ElementType.METHOD,
-        ElementType.TYPE, ElementType.ANNOTATION_TYPE})
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-@Qualifier("redisCacheConnectionFactory")
-public @interface DomainRedisConnectionFactory {
-}
