--- pennyway-domain/src/main/java/kr/co/pennyway/common/converter/LegacyCommonType.java
@@ -1,10 +1,0 @@
-package kr.co.pennyway.common.converter;
-
-public interface LegacyCommonType {
-    /**
-     * Legacy Super System 공통 코드를 반환한다.
-     *
-     * @return String 공통 코드
-     */
-    String getCode();
-}
