--- pennyway-domain/src/main/java/kr/co/pennyway/config/JpaConfig.java
@@ -1,14 +1,0 @@
-package kr.co.pennyway.config;
-
-import kr.co.pennyway.DomainPackageLocation;
-import org.springframework.boot.autoconfigure.domain.EntityScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-
-@Configuration
-@EnableJpaAuditing
-@EntityScan(basePackageClasses = DomainPackageLocation.class)
-@EnableJpaRepositories(basePackageClasses = DomainPackageLocation.class)
-public class JpaConfig {
-}
