+++ pennyway-domain/src/main/java/kr/co/pennyway/domain/DomainPackageLocation.java
@@ -1,0 +1,4 @@
+package kr.co.pennyway.domain;
+
+public interface DomainPackageLocation {
+}
