+++ pennyway-domain/src/main/java/kr/co/pennyway/domain/common/annotation/DomainRedisTemplate.java
@@ -1,0 +1,13 @@
+package kr.co.pennyway.domain.common.annotation;
+
+import org.springframework.beans.factory.annotation.Qualifier;
+
+import java.lang.annotation.*;
+
+@Target({ElementType.FIELD, ElementType.PARAMETER, ElementType.METHOD,
+        ElementType.TYPE, ElementType.ANNOTATION_TYPE})
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+@Qualifier("domainRedisTemplate")
+public @interface DomainRedisTemplate {
+}
