+++ pennyway-domain/src/main/java/kr/co/pennyway/domain/common/converter/LegacyCommonType.java
@@ -1,0 +1,10 @@
+package kr.co.pennyway.domain.common.converter;
+
+public interface LegacyCommonType {
+    /**
+     * Legacy Super System 공통 코드를 반환한다.
+     *
+     * @return String 공통 코드
+     */
+    String getCode();
+}
